Fix integer type required for QT widget (float was passed).

+++ uiutils.py
@@ -129,7 +129,7 @@ def jsonFont(fnt, scale):
     if font_italic == '1':
         fnt_flags |= gui.QFont.StyleItalic
 
-    font_size = font_size / float(scale) * 14.
+    font_size = int(font_size / float(scale) * 14.)
 
     qfnt = gui.QFont(font_name, font_size, fnt_flags);
     qfnt.setPixelSize(font_size)
